+++ tests/unit/test_utils.h
@@ -17,7 +17,7 @@ void compareFlatIndexInfoToIterator(VecSimIndexInfo info, VecSimInfoIterator *in
 
 void compareHNSWIndexInfoToIterator(VecSimIndexInfo info, VecSimInfoIterator *infoIter);
 
-void runRangeQueryTest(VecSimIndex *index, const void *query, float radius,
+void runRangeQueryTest(VecSimIndex *index, const void *query, double radius,
                        const std::function<void(size_t, double, size_t)> &ResCB,
                        size_t expected_res_num, VecSimQueryResult_Order order = BY_ID,
                        VecSimQueryParams *params = nullptr);
